Fix button labels in dialog
The dialog to enter a name for the compressed file is based in a
standard "OC.dialogs.prompt", but with some adjustments, like the button
labels. The adjustment relies on the whole dialog containing only two
buttons (those from the button row), which was true in the past.
However, due to the accesibility fixes the dialog now contains an extra
button to close the dialog, which alters the expected order of the
buttons to modify and causes the labels to be wrongly assigned.

To fix that the query to select the buttons to modify is restricted only
to those in the button row, which behaves as expected in the whole range
of supported versions (from Nextcloud 22 to Nextcloud 27).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>