feat: Add "Compress to zip" action with declarative UI
A new endpoint was added to compress a given path to zip. A path had to
be used instead of a file ID as the same file ID could belong to
different paths, so just from the file ID it would not be possible to
generate the path to the target zip file.

A unique path will be generated for the target zip file when the zip job
is created. However, note that creating the zip file will anyway fail if
the path exists once the job is executed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>